feat: Moving the ChatBot
In file chatlogic.cpp, create a local ChatBot instance on the stack
at the bottom of function LoadAnswerGraphFromFile. Then, pass the
ChatBot into the root node using move semantics. Ensure that CharLogic
has no ownership relation to the ChatBot instance and thus is no
longer responsible for memory allocation and deallocation.

Note that the member _chatBot of ChatLogic remains so it can be used
as a communication handle between GUI and ChatBot instance. Required
changes are made to files chatlogic.h / chatlogic.cpp and
graphnode.h / graphnode.cpp.

When the program is executed, messages on which part of the
Rule of Five components of ChatBot is called will be printed to
the console. When sending a query to the ChatBot, the output will
look like the following:

ChatBot Constructor
ChatBot Move Constructor
ChatBot Move Assignment Operator
ChatBot Destructor
ChatBot Destructor